Move useRouter to lib
