+++ frontend/src/components/NavBar.js
@@ -18,7 +18,7 @@ import { savingIndicator } from '../apollo/apolloClient'
 import { Role } from '../lib/permissions'
 import { PROJECT_BY_ID, WORKSPACE_BY_ID, COURSE_BY_ID } from '../graphql/Query'
 import { useLoginStateValue } from '../store'
-import useRouter from '../useRouter'
+import useRouter from '../lib/useRouter'
 import { useLoadingBar } from './LoadingBar'
 
 const Link = props => <MaterialLink {...props} component={RouterLink} />
