+++ frontend/src/views/login/LoginView.js
@@ -9,7 +9,7 @@ import qs from 'qs'
 import { MERGE_USER } from '../../graphql/Mutation'
 import Auth from '../../lib/authentication'
 import { useLoginStateValue, useMessageStateValue } from '../../store'
-import useRouter from '../../useRouter'
+import useRouter from '../../lib/useRouter'
 import { ReactComponent as HakaIcon } from '../../static/haka.svg'
 import { noDefault } from '../../lib/eventMiddleware'
 import LoadingBar from '../../components/LoadingBar'
