+++ frontend/src/views/user/UserView.js
@@ -6,7 +6,7 @@ import { useMutation } from 'react-apollo-hooks'
 import { useLoginStateValue, useMessageStateValue } from '../../store'
 import { Role } from '../../lib/permissions'
 import Auth from '../../lib/authentication'
-import useRouter from '../../useRouter'
+import useRouter from '../../lib/useRouter'
 import { DISCONNECT_AUTH, MERGE_USER } from '../../graphql/Mutation'
 import { PROJECTS_FOR_USER, WORKSPACES_FOR_USER } from '../../graphql/Query'
 import { useLoginDialog } from '../../dialogs/authentication'
